+++ code/src/sixsq/nuvla/server/resources/nuvlabox/status_utils.clj
@@ -1,7 +1,9 @@
 (ns sixsq.nuvla.server.resources.nuvlabox.status-utils
   (:require
     [sixsq.nuvla.server.resources.common.crud :as crud]
-    [sixsq.nuvla.server.util.time :as time]))
+    [sixsq.nuvla.server.util.time :as time]
+    [sixsq.nuvla.auth.utils :as auth]
+    [clojure.string :as str]))
 
 
 (defn get-next-heartbeat
@@ -14,3 +16,11 @@
             time/to-str)
     (catch Exception _
       nil)))
+
+
+(defn set-online
+  [resource request]
+  (let [active-claim (auth/current-active-claim request)
+        is-nuvlabox? (str/starts-with? active-claim "nuvlabox/")]
+    (cond-> resource
+            is-nuvlabox? (assoc :online true))))