+++ code/src/sixsq/nuvla/server/resources/nuvlabox_status.clj
@@ -15,6 +15,7 @@ Versioned subclasses define the attributes for a particular NuvlaBox release.
     [sixsq.nuvla.server.resources.resource-metadata :as md]
     [sixsq.nuvla.server.resources.spec.nuvlabox-status :as nb-status]
     [sixsq.nuvla.server.util.metadata :as gen-md]
+    [sixsq.nuvla.server.resources.nuvlabox.status-utils :as status-utils]
     [sixsq.nuvla.server.util.response :as r]))
 
 
@@ -110,6 +111,7 @@ Versioned subclasses define the attributes for a particular NuvlaBox release.
       (-> merged
           (u/update-timestamps)
           (u/set-updated-by request)
+          (status-utils/set-online request)
           pre-edit
           crud/validate
           (db/edit request)))